Fix all the merge conflicts
